+++ inc/extra.h
@@ -6,7 +6,7 @@
 /*   By: aquinter <aquinter@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/07 12:59:00 by lgandari          #+#    #+#             */
-/*   Updated: 2024/07/08 16:50:56 by aquinter         ###   ########.fr       */
+/*   Updated: 2024/07/09 18:27:24 by lgandari         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,6 @@
 void	free_matrix(char **matrix);
 int		ft_strcmp(const char *s1, const char *s2);
 int		print_error(char *msg);
-int		ft_arrlen(void **arr);
+size_t	ft_arrlen(void **arr);
 
 #endif
